backend/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "knowledge-platform-backend"
version = "0.1.0"
description = "Backend for Knowledge Discovery Platform"
requires-python = ">=3.12"
authors = [
    {name = "Knowledge Platform Team"}
]
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.29.0",
    "alembic>=1.13.0",
    "httpx>=0.26.0",
    "pyjwt>=2.8.0",
    "python-jose[cryptography]>=3.3.0",
    "bleach>=6.1.0",
    "pydantic-settings>=2.1.0",
    "python-multipart>=0.0.6",
    "ruff>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["app/tests"]
